requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
psycopg2-binary
python-dotenv